Enable TCP Keep-Alives
As referenced in https://github.com/contribsys/faktory/issues/273#issuecomment-576463404, TCP Keep-Alives are used to keep a connection marked as active. While neither the worker nor heartbeat threads have a moment in which the connection goes quiet, we've seen behaviour matching the original issue.

This changes enables TCP Keep-Alives as a precaution.